+++ FALOFinancialProofing/Models/Campaign.cs
@@ -23,7 +23,7 @@ namespace FALOFinancialProofing.Models
 
         [Column(TypeName = "money")]
         [Required]
-        public double FundTarget { get; set; }
+        public long FundTarget { get; set; }
 
         [MaxLength(200)]
         public string? Image { get; set; }
